// frontend/src/api/models/createManualAccountDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { CreateManualAccountDtoType } from './createManualAccountDtoType'
import type { CreateManualAccountDtoSubType } from './createManualAccountDtoSubType'
import type { MoneyWithSign } from './moneyWithSign'

export interface CreateManualAccountDto {
  /** @nullable */
  name: string | null
  /** @nullable */
  customName?: string | null
  /** @nullable */
  notes?: string | null
  type: (typeof CreateManualAccountDtoType)[keyof typeof CreateManualAccountDtoType]
  /** @nullable */
  subType: CreateManualAccountDtoSubType
  availableBalance: MoneyWithSign
  currentBalance: MoneyWithSign
}