// frontend/src/api/models/createManualTransactionDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MoneyWithSign } from './moneyWithSign'

export interface CreateManualTransactionDto {
  accountId: string
  amount: MoneyWithSign
  /** @minLength 1 */
  merchantName: string
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  providerDate: string
  categoryId: string
}